% Auto-generated: do not edit by hand
\name{dfcBackground}

\alias{dfcBackground}

\title{Background component}

\description{

}

\usage{
dfcBackground(id=NULL, className=NULL, color=NULL, gap=NULL, size=NULL,
style=NULL, variant=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. }

\item{color}{Character. }

\item{gap}{Numeric. }

\item{size}{Numeric. }

\item{style}{Named list. }

\item{variant}{Character. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

