% Auto-generated: do not edit by hand
\name{dfcControls}

\alias{dfcControls}

\title{Controls component}

\description{

}

\usage{
dfcControls(id=NULL, className=NULL, onFitView=NULL,
onInteractiveChange=NULL, onZoomIn=NULL, onZoomOut=NULL,
showFitView=NULL, showInteractive=NULL, showZoom=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{className}{Character. }

\item{onFitView}{}

\item{onInteractiveChange}{}

\item{onZoomIn}{Logical. }

\item{onZoomOut}{}

\item{showFitView}{Logical. }

\item{showInteractive}{Logical. }

\item{showZoom}{Logical. }

\item{style}{Named list. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

