% Auto-generated: do not edit by hand
\name{dfcFlow}

\alias{dfcFlow}

\title{Flow component}

\description{

}

\usage{
dfcFlow(id=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

