% Auto-generated: do not edit by hand
\name{dfcFlow}

\alias{dfcFlow}

\title{Flow component}

\description{

}

\usage{
dfcFlow(children=NULL, id=NULL, elements=NULL, snapGrid=NULL,
style=NULL)
}

\arguments{
\item{children}{Logical | numeric | character | named list | unnamed list. }

\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{elements}{List of named lists. Node/edge elements contained within svg space.}

\item{snapGrid}{List of numerics. }

\item{style}{Named list. CSS style attributes for div wrapper around ReactFlow component.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

